Refactor: Core architecture improvements
- Split database.py into connection, query_executor, and lifecycle modules
- Reorganize schemas package with mixins and clear separation of concerns
- Decompose SessionRepository into ProgressRepository and LifecycleStateRepository
- Update all unit tests to reflect architectural changes
- Remove deprecated files and fix circular dependencies
