+++ src/state_db/repositories/__init__.py
@@ -1,11 +1,15 @@
 from state_db.repositories.entity import EntityRepository
+from state_db.repositories.lifecycle_state import LifecycleStateRepository
 from state_db.repositories.player import PlayerRepository
+from state_db.repositories.progress import ProgressRepository
 from state_db.repositories.scenario import ScenarioRepository
 from state_db.repositories.session import SessionRepository
 from state_db.repositories.trace import TraceRepository
 
 __all__ = [
     "SessionRepository",
+    "ProgressRepository",
+    "LifecycleStateRepository",
     "PlayerRepository",
     "EntityRepository",
     "TraceRepository",
