+++ src/state_db/routers/router_INQUIRY.py
@@ -17,7 +17,9 @@ from state_db.models import (
 )
 from state_db.repositories import (
     EntityRepository,
+    LifecycleStateRepository,
     PlayerRepository,
+    ProgressRepository,
     ScenarioRepository,
     SessionRepository,
 )
@@ -25,7 +27,9 @@ from state_db.schemas import ScenarioInfo
 
 from .dependencies import (
     get_entity_repo,
+    get_lifecycle_repo,
     get_player_repo,
+    get_progress_repo,
     get_scenario_repo,
     get_session_repo,
 )
@@ -157,7 +161,8 @@ async def get_enemies(
     "/session/{session_id}/phase", response_model=WrappedResponse[PhaseChangeResult]
 )
 async def get_phase(
-    session_id: str, repo: Annotated[SessionRepository, Depends(get_session_repo)]
+    session_id: str,
+    repo: Annotated[LifecycleStateRepository, Depends(get_lifecycle_repo)],
 ) -> Dict[str, Any]:
     result = await repo.get_phase(session_id)
     return {"status": "success", "data": result}
@@ -165,7 +170,8 @@ async def get_phase(
 
 @router.get("/session/{session_id}/turn", response_model=WrappedResponse[TurnAddResult])
 async def get_turn(
-    session_id: str, repo: Annotated[SessionRepository, Depends(get_session_repo)]
+    session_id: str,
+    repo: Annotated[LifecycleStateRepository, Depends(get_lifecycle_repo)],
 ) -> Dict[str, Any]:
     result = await repo.get_turn(session_id)
     return {"status": "success", "data": result}
@@ -178,7 +184,7 @@ async def get_turn(
 
 @router.get("/session/{session_id}/act", response_model=WrappedResponse[Dict[str, Any]])
 async def get_act(
-    session_id: str, repo: Annotated[SessionRepository, Depends(get_session_repo)]
+    session_id: str, repo: Annotated[ProgressRepository, Depends(get_progress_repo)]
 ) -> Dict[str, Any]:
     result = await repo.get_act(session_id)
     return {"status": "success", "data": result}
@@ -199,7 +205,7 @@ async def get_current_act_details(
     "/session/{session_id}/sequence", response_model=WrappedResponse[Dict[str, Any]]
 )
 async def get_sequence(
-    session_id: str, repo: Annotated[SessionRepository, Depends(get_session_repo)]
+    session_id: str, repo: Annotated[ProgressRepository, Depends(get_progress_repo)]
 ) -> Dict[str, Any]:
     result = await repo.get_sequence(session_id)
     return {"status": "success", "data": result}
@@ -209,7 +215,7 @@ async def get_sequence(
     "/session/{session_id}/location", response_model=WrappedResponse[Dict[str, Any]]
 )
 async def get_location(
-    session_id: str, repo: Annotated[SessionRepository, Depends(get_session_repo)]
+    session_id: str, repo: Annotated[ProgressRepository, Depends(get_progress_repo)]
 ) -> Dict[str, Any]:
     result = await repo.get_location(session_id)
     return {"status": "success", "data": result}
@@ -219,7 +225,7 @@ async def get_location(
     "/session/{session_id}/progress", response_model=WrappedResponse[Dict[str, Any]]
 )
 async def get_progress(
-    session_id: str, repo: Annotated[SessionRepository, Depends(get_session_repo)]
+    session_id: str, repo: Annotated[ProgressRepository, Depends(get_progress_repo)]
 ) -> Dict[str, Any]:
     result = await repo.get_progress(session_id)
     return {"status": "success", "data": result}
