+++ src/state_db/schemas/__init__.py
@@ -1,23 +1,26 @@
-from .base import (
-    EnemyBase,
-    EntityBase,
-    ItemBase,
-    NPCBase,
-    Phase,
-    PlayerBase,
-    SessionBase,
-    StateBase,
+from .base_entities import EnemyBase, ItemBase, NPCBase, PlayerBase
+from .management import (
+    SessionControlResponse,
+    SessionInfoResponse,
+    SessionStartRequest,
+    SessionStartResponse,
 )
-from .enemy import EnemyHPUpdateRequest, EnemySpawnRequest
-from .inventory import InventoryItem, InventoryUpdateRequest, InventoryUpdateResponse
-from .item import ItemEarnRequest, ItemInfoResponse, ItemUseRequest
-from .npc import NPCAffinityUpdateRequest, NPCSpawnRequest
-from .player import (
-    NPCRelation,
-    PlayerData,
+from .management_requests import (
+    ActChangeRequest,
+    EnemySpawnRequest,
+    NPCSpawnRequest,
+    PhaseChangeRequest,
+    SequenceChangeRequest,
+)
+from .mixins import EntityBaseMixin, SessionContextMixin, StateMixin
+from .requests import (
+    EnemyHPUpdateRequest,
+    InventoryUpdateRequest,
+    ItemEarnRequest,
+    ItemUseRequest,
+    LocationUpdateRequest,
+    NPCAffinityUpdateRequest,
     PlayerHPUpdateRequest,
-    PlayerStateRequest,
-    PlayerStateResponse,
     PlayerStatsUpdateRequest,
 )
 from .scenario import (
@@ -31,56 +34,33 @@ from .scenario import (
     ScenarioInjectResponse,
     ScenarioSequenceInject,
 )
-from .session import (
-    SessionEndResponse,
-    SessionInfoResponse,
-    SessionPauseResponse,
-    SessionResumeResponse,
-    SessionStartRequest,
-    SessionStartResponse,
-)
-from .world import (
-    ActChangeRequest,
-    LocationUpdateRequest,
-    PhaseChangeRequest,
-    SequenceChangeRequest,
-)
+from .system import Phase, TurnRecord
 
 __all__ = [
-    "Phase",
-    "SessionBase",
-    "EntityBase",
-    "StateBase",
+    # Mixins
+    "SessionContextMixin",
+    "EntityBaseMixin",
+    "StateMixin",
+    # Base Entities
     "PlayerBase",
     "NPCBase",
     "EnemyBase",
     "ItemBase",
+    # System
+    "Phase",
+    "TurnRecord",
+    # Management
     "SessionStartRequest",
     "SessionStartResponse",
-    "SessionEndResponse",
-    "SessionPauseResponse",
-    "SessionResumeResponse",
+    "SessionControlResponse",
     "SessionInfoResponse",
-    "InventoryUpdateRequest",
-    "InventoryItem",
-    "InventoryUpdateResponse",
-    "ItemInfoResponse",
-    "ItemEarnRequest",
-    "ItemUseRequest",
-    "PlayerStateRequest",
-    "PlayerData",
-    "NPCRelation",
-    "PlayerStateResponse",
-    "PlayerHPUpdateRequest",
-    "PlayerStatsUpdateRequest",
-    "NPCAffinityUpdateRequest",
-    "NPCSpawnRequest",
-    "EnemySpawnRequest",
-    "EnemyHPUpdateRequest",
-    "LocationUpdateRequest",
-    "PhaseChangeRequest",
+    # Management Requests
     "ActChangeRequest",
     "SequenceChangeRequest",
+    "PhaseChangeRequest",
+    "EnemySpawnRequest",
+    "NPCSpawnRequest",
+    # Scenario
     "ScenarioActInject",
     "ScenarioSequenceInject",
     "ScenarioInjectNPC",
@@ -90,4 +70,13 @@ __all__ = [
     "ScenarioInjectRequest",
     "ScenarioInjectResponse",
     "ScenarioInfo",
+    # Requests (Update)
+    "PlayerHPUpdateRequest",
+    "PlayerStatsUpdateRequest",
+    "InventoryUpdateRequest",
+    "NPCAffinityUpdateRequest",
+    "LocationUpdateRequest",
+    "EnemyHPUpdateRequest",
+    "ItemEarnRequest",
+    "ItemUseRequest",
 ]
